webscraping_summarizer/requirements.txt
streamlit==1.28.1
python-dotenv==1.0.0
newspaper3k==0.2.8
langchain==0.0.350
langchain-groq==0.0.1 